+++ index.ts
@@ -1,3 +1,4 @@
+import { Configuration } from './core/Configuration';
 import {Registrar} from './core/Registrar';
 import {IocContainer} from './core/IoC';
 import {Router} from 'express';
@@ -10,6 +11,7 @@ export * from './errors/Errors';
 export * from './params/ParamDecorators';
 export * from './routes/RouteDecorators';
 export * from './validators/Validators';
+export * from './versioning/VersionDecorator';
 
 let registrar = IocContainer.get<Registrar>(IoCSymbols.registrar);
 
@@ -36,3 +38,9 @@ export const registerControllersFromFolder = registrar.registerControllersFromFo
  * @returns {Router} - The configured router.
  */
 export const registerControllers = registrar.registerControllers.bind(registrar);
+
+/**
+ * Injected object that contains the global configuration for giuseppe.
+ * May be changed before the registration process (e.g. accept header version setting).
+ */
+export const giuseppeConfiguration = IocContainer.get<Configuration>(IoCSymbols.configuration);