+++ CHANGELOG.md
@@ -3,10 +3,15 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Versioning for routes and controllers.
+
 ### Changed
 - Moved documentation to gh-pages.
 - Use `chai-http` instead of `newman` for integration tests.
 - Core rewrite (some IoC things, moved classes, etc).
+- Upgraded to TypeScript 2.0
+- Remap lcov coverage back to typescript
 
 ## [1.1.1]
 ### Fixed